requirements.txt
fastapi
uvicorn
dash
dash-auth
pandas
numpy
scikit-learn
plotly
ta
websockets
requests
pyyaml
flask
technical-analysis
